fix(agent): render subagent tool failures + harden memory lock symlinks
Follow-up to the #4504 review:
- Extract ToolFailureLines into a shared component and render it in the
  sub-agent renderers (SubagentDrawer rows + ToolTimelineBlock inline rows), so
  a failed child tool actually SHOWS the why/next copy instead of only storing
  it on the transcript item (#4459).
- update_memory_md: reject a symlinked `.memory-write.lock` and open it
  O_NOFOLLOW on Unix, so a project-controlled symlink can't redirect the
  cross-process flock outside the containment-checked workspace (#4458).
